Added DatabaseButton component
